Fixed crash from 8:00-9:59.
date %H outputs 2 digits (always).
bash treats anything starting with a 0 as octal.
This works fine with 2 digits or a number < 8,
however, at 8:00, date %H outputs 08, which makes bash think it is an
octal number, but 8 isn't a valid octal digit, so it returns false.
This causes night mode to trigger.
Workaround: sed away the 0.